+++ scrapper/src/main/java/edu/java/service/LinkUpdaterScheduler.java
@@ -10,9 +10,7 @@ import java.util.List;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Service;
 
-@Service
 @Slf4j
 public class LinkUpdaterScheduler {
